+++ app/src/main/kotlin/com/simplemobiletools/contacts/helpers/VcfExporter.kt
@@ -6,11 +6,9 @@ import android.provider.ContactsContract.CommonDataKinds
 import android.provider.MediaStore
 import android.util.Base64
 import com.simplemobiletools.commons.activities.BaseSimpleActivity
-import com.simplemobiletools.commons.extensions.getFileOutputStream
-import com.simplemobiletools.commons.extensions.showErrorToast
-import com.simplemobiletools.commons.extensions.toFileDirItem
-import com.simplemobiletools.commons.extensions.writeLn
-import com.simplemobiletools.contacts.helpers.VcfExporter.ExportResult.*
+import com.simplemobiletools.commons.extensions.*
+import com.simplemobiletools.contacts.R
+import com.simplemobiletools.contacts.helpers.VcfExporter.ExportResult.EXPORT_FAIL
 import com.simplemobiletools.contacts.models.Contact
 import java.io.BufferedWriter
 import java.io.ByteArrayOutputStream
@@ -26,12 +24,16 @@ class VcfExporter {
     private var contactsExported = 0
     private var contactsFailed = 0
 
-    fun exportContacts(activity: BaseSimpleActivity, file: File, contacts: ArrayList<Contact>, callback: (result: ExportResult) -> Unit) {
-        try {
-            activity.getFileOutputStream(file.toFileDirItem(activity)) {
+    fun exportContacts(activity: BaseSimpleActivity, file: File, contacts: ArrayList<Contact>, showExportingToast: Boolean, callback: (result: ExportResult) -> Unit) {
+        activity.getFileOutputStream(file.toFileDirItem(activity), true) {
+            try {
                 if (it == null) {
                     callback(EXPORT_FAIL)
                     return@getFileOutputStream
+                }
+
+                if (showExportingToast) {
+                    activity.toast(R.string.exporting)
                 }
 
                 it.bufferedWriter().use { out ->
@@ -53,7 +55,7 @@ class VcfExporter {
                         contact.addresses.forEach {
                             val type = getAddressTypeLabel(it.type)
                             val delimiterType = if (type.isEmpty()) "" else ";$type"
-                            out.writeLn("$ADR$delimiterType:;;${it.value};;;;")
+                            out.writeLn("$ADR$delimiterType:;;${it.value.replace("\n", "\\n")};;;;")
                         }
 
                         contact.events.forEach {
@@ -71,6 +73,10 @@ class VcfExporter {
                             out.writeLn("$TITLE${contact.organization.jobPosition.replace("\n", "\\n")}")
                         }
 
+                        contact.websites.forEach {
+                            out.writeLn("$URL$it")
+                        }
+
                         if (contact.thumbnailUri.isNotEmpty()) {
                             val bitmap = MediaStore.Images.Media.getBitmap(activity.contentResolver, Uri.parse(contact.thumbnailUri))
                             addBitmap(bitmap, out)
@@ -84,16 +90,17 @@ class VcfExporter {
                         contactsExported++
                     }
                 }
-            }
-        } catch (e: Exception) {
-            activity.showErrorToast(e)
-        }
 
-        callback(when {
-            contactsExported == 0 -> EXPORT_FAIL
-            contactsFailed > 0 -> EXPORT_PARTIAL
-            else -> EXPORT_OK
-        })
+            } catch (e: Exception) {
+                activity.showErrorToast(e)
+            }
+
+            callback(when {
+                contactsExported == 0 -> EXPORT_FAIL
+                contactsFailed > 0 -> ExportResult.EXPORT_PARTIAL
+                else -> ExportResult.EXPORT_OK
+            })
+        }
     }
 
     private fun addBitmap(bitmap: Bitmap, out: BufferedWriter) {